Refactor kata based on three intern interviews
This is a lot to throw at someone that has never used clojure
before. If you have a one-hour interview slot and are comfortable
pairing with someone junior, this is one heck of a way to see
how someone responds to discomfort and an unknown codebase and
language.

I strongly recommend articulating repeatedly that you expect
this to be difficult.

I would avoid going through the README entirely, that's a waste
of time during a short interview.